requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
aiohttp==3.8.5
asgiref==3.7.2
uvicorn==0.23.2
nltk==3.8.1
scikit-learn==1.3.0
async-timeout==4.0.3
aiodns==3.0.0
tenacity==8.2.2
